refactor: move classes to separate files
